zoom factor is float, not int

+++ avidemux/qt4/ADM_render_qt/GUI_render.h
@@ -83,7 +83,7 @@ typedef struct
 {
     int phyW;
     int phyH;
-    int lastZoom;
+    float lastZoom;
     GUI_WindowInfo xinfo;
     void *draw;
 } ADM_renderContext;
